suggestions not index, java big integer

+++ src/main/scala/com/travel_stories/database/Place.scala
@@ -1,14 +1,15 @@
 package com.travel_stories.database
+import java.math.BigInteger
 
 
-class Place(key:BigInt, name:String, latitude:Double, longitude:Double){
+class Place(key:BigInteger, name:String, latitude:Double, longitude:Double){
   
   var popularity:Int = 0
   
   def setPopularity(p:Int) = popularity = p
   
   def getName():String = name
-  def getKey():BigInt = key
+  def getKey():BigInteger = key
   def getLatitude():Double = latitude
   def getLongitude():Double = longitude
   def getPopularity():Int = popularity
